Homogenizing all meshes to double for simplicity

+++ files/ionprinter/simulation/nyion/src/data_structure.hpp
@@ -4,7 +4,7 @@
 class root_mesh_geometry{
   public:
 
-    root_mesh_geometry(float bounds[6], float root_scale);
+    root_mesh_geometry(double bounds[6], double root_scale);
 
     int root_x_len; //root gridpoints for the mesh
     int root_y_len;
@@ -12,14 +12,14 @@ class root_mesh_geometry{
 
     int root_size; //total gridpoints in root
 
-    float root_scale; //world scale to submesh cube scale
+    double root_scale; //world scale to submesh cube scale
 
-    float x_min_bound;
-    float x_max_bound;
-    float y_min_bound;
-    float y_max_bound;
-    float z_min_bound;
-    float z_max_bound; //f*** getters and setters.
+    double x_min_bound;
+    double x_max_bound;
+    double y_min_bound;
+    double y_max_bound;
+    double z_min_bound;
+    double z_max_bound; //f*** getters and setters.
 
 };
 //
